requirements_dev.txt
redis
python-stdnet
oauthlib
